requirements.txt
adminlte-base==0.2.0
arrow==0.17.0
asgiref==3.3.1
Babel==2.9.0
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-adminlte-full==0.2.0
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-phone-login==1.1.3
django-phonenumber-field==1.1.0
django-sendsms==0.2.3
djangorestframework==3.12.2
gunicorn==20.0.4
idna==2.10
Jinja2==2.11.2
MarkupSafe==1.1.1
phonenumberslite==8.12.14
psycopg2-binary==2.8.6
pygraphviz==1.6
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
vobject==0.9.6.1
